Slides now take in content as props
